requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
databases==0.5.3
aiosqlite==0.17.0
python-multipart==0.0.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
jinja2==3.0.1
python-dotenv==0.19.0
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
python-dateutil==2.8.2
openpyxl==3.0.9
Flask-Login==0.5.0
werkzeug==2.0.1
gunicorn==20.1.0
Flask-Limiter==2.1